feat: setup Xcode project with OpenCV integration
- Generate Xcode project using xcodegen
- Integrate OpenCV 4.13 via Homebrew dylib linking
- Fix macOS 12.0+ API compatibility issues:
  - SCShareableContent requires macOS 12.3+
  - SCScreenshotManager requires macOS 14.0+
  - canJoinAllApplications requires macOS 13.0+
  - formStyle(.grouped) requires macOS 13.0+
- Fix OpenCV/Apple NO macro conflict in ObjC++
- Add ImageIO import for CGImageDestination functions
- Fix Int to Int32 type conversions for ObjC bridge


AI-Model: glm-5
AI-Contributed/Feature: 319/900
AI-Contributed/UT: 0/0

+++ ScrollCapture/Engine/StitchEngine.swift
@@ -23,9 +23,9 @@ class StitchEngine {
 
     private init() {
         bridge = StitchEngineBridge(
-            templateHeight: Constants.Defaults.templateHeight,
+            templateHeight: Int32(Constants.Defaults.templateHeight),
             matchThreshold: Constants.Defaults.matchThreshold,
-            maxResultHeight: Constants.Defaults.maxResultHeight
+            maxResultHeight: Int32(Constants.Defaults.maxResultHeight)
         )
     }
